compute density on hedgeData save
